Add split_tracks
This function allows to split a single track with a large gap in the middle (e.g. if there is a 2.5 minute gap in a single track) into separate tracks. This came in reaction to some tracks that were obviously made up from multiple cells that were erroneously lumped together. We've also made changes to the upstream track inference, so this isn't currently used.